str-format: fixed a compilation warning

+++ lib/str-format.c
@@ -87,7 +87,7 @@ scan_uint32(const gchar **buf, gint *left, gint field_width, guint32 *num)
     {
       if ((**buf) >= '0' && (**buf) <= '9')
         result = result * 10 + ((**buf) - '0');
-      else if (!isspace(**buf))
+      else if (!isspace((int) **buf))
         return FALSE;
       (*buf)++;
       (*left)--;
